requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-role-permissions==3.2.0
pillow==11.1.0
sqlparse==0.5.3
